+++ spec/lib/karafka/connection/consumer_spec.rb
@@ -184,7 +184,7 @@ RSpec.describe Karafka::Connection::Consumer do
             logger: ::Karafka.logger,
             client_id: ::Karafka::App.config.client_id,
             seed_brokers: ::Karafka::App.config.kafka.seed_brokers,
-            socket_timeout: 10,
+            socket_timeout: 30,
             connect_timeout: 10,
             sasl_plain_authzid: ''
           )